Moved declaration of Mover and Object type enums
