+++ BumbleBots/Objects.h
@@ -12,6 +12,23 @@
 #include "Utils.h"
 
 //-----------------------------------------------------------------------------
+// ObjectType declaration
+
+enum class ObjectType : int8_t {
+  Pickup,
+  Teleport,
+  Gap,
+  Obstacle,
+  MenuDigit
+};
+
+const ObjectType TYPE_PICKUP = ObjectType::Pickup;
+const ObjectType TYPE_TELEPORT = ObjectType::Teleport;
+const ObjectType TYPE_GAP = ObjectType::Gap;
+const ObjectType TYPE_OBSTACLE = ObjectType::Obstacle;
+const ObjectType TYPE_MENUDIGIT = ObjectType::MenuDigit;
+
+//-----------------------------------------------------------------------------
 // Object declaration
 
 class Object {
