+++ BumbleBots/Utils.h
@@ -19,28 +19,6 @@ const int8_t NORTH_EAST = 0;
 //const int8_t SOUTH_WEST = 2;
 //const int8_t NORTH_WEST = 3;
 
-enum class MoverType : int8_t {
-  Player,
-  Enemy,
-  Box
-};
-const MoverType TYPE_PLAYER = MoverType::Player;
-const MoverType TYPE_ENEMY = MoverType::Enemy;
-const MoverType TYPE_BOX = MoverType::Box;
-
-enum class ObjectType : int8_t {
-  Pickup,
-  Teleport,
-  Gap,
-  Obstacle,
-  MenuDigit
-};
-const ObjectType TYPE_PICKUP = ObjectType::Pickup;
-const ObjectType TYPE_TELEPORT = ObjectType::Teleport;
-const ObjectType TYPE_GAP = ObjectType::Gap;
-const ObjectType TYPE_OBSTACLE = ObjectType::Obstacle;
-const ObjectType TYPE_MENUDIGIT = ObjectType::MenuDigit;
-
 /* Valid values: 0..maxCols * maxRows - 1
  * Using signed to enable more efficient iteration (and support for off-map positions)
  */
